grub: update to latest (2.06 -> 2.12)
Not simply a rebase, the changes are incompatible with older versions of
TrenchBoot Linux and SKL.

Bash completion seems to get installed now.

Also work around a workaround in poky:

Tarballs of 2.12 are missing an automatically created file, so poky
creates it not taking into account that this breaks out-of-tree builds:
Make finds the file as "extra_deps.lst" in the source tree for its
dependency checking through VPATH, but shell commands run in the build
tree don't check the source tree and just fail.

Signed-off-by: Sergii Dmytruk <sergii.dmytruk@3mdeb.com>